refactor(api): move missing tests
