+++ api/tests/shared/unit/application/badges/index_test.js
@@ -1,7 +1,7 @@
-import { expect, sinon, HttpTestServer } from '../../../test-helper.js';
-import { badgesController } from '../../../../src/shared/application/badges/badges-controller.js';
-import { securityPreHandlers } from '../../../../lib/application/security-pre-handlers.js';
-import * as badgesRouter from '../../../../src/shared/application/badges/index.js';
+import { expect, sinon, HttpTestServer } from '../../../../test-helper.js';
+import { badgesController } from '../../../../../src/shared/application/badges/badges-controller.js';
+import { securityPreHandlers } from '../../../../../lib/application/security-pre-handlers.js';
+import * as badgesRouter from '../../../../../src/shared/application/badges/index.js';
 
 describe('Unit | Application | Badges | Routes', function () {
   describe('DELETE /api/admin/badges/{id}', function () {
